add purchase functionality

+++ src/main/java/com/example/eshop/DTO/ItemCreateDTO.java
@@ -6,7 +6,7 @@ public record ItemCreateDTO(
         String name,
         String brand,
         String description,
-        BigDecimal price,
+        double price,
         int quantity
 ) {
 }
